auto-pc-miner/requirements.txt
pywin32
python-dotenv